refactoring timeFormat to be used as util
