+++ src/js/components/Video/video.stories.js
@@ -26,6 +26,26 @@ const SimpleVideo = () => (
         />
       </Video>
     </Box>
+    <Box align="center" pad="large">
+      <video controls>
+        <source
+          src="http://techslides.com/demos/sample-videos/small.webm"
+          type="video/webm"
+        />
+        <source
+          src="http://techslides.com/demos/sample-videos/small.ogv"
+          type="video/ogg"
+        />
+        <source
+          src="http://techslides.com/demos/sample-videos/small.mp4"
+          type="video/mp4"
+        />
+        <source
+          src="http://techslides.com/demos/sample-videos/small.3gp"
+          type="video/3gp"
+        />
+      </video>
+    </Box>
   </Grommet>
 );
