+++ src/js/utils/index.js
@@ -1,12 +1,14 @@
-export * from './mixins';
+
 export * from './background';
 export * from './colors';
 export * from './debounce';
 export * from './DOM';
 export * from './graphics';
-export * from './prop-types';
-export * from './styles';
+export * from './mixins';
 export * from './object';
+export * from './prop-types';
 export * from './responsive';
 export * from './router';
+export * from './styles';
+export * from './time';
 export * from './throttle';
